requirements.txt
ninja  # For faster builds.
psutil
ray >= 2.5.1
pandas  # Required for Ray data.
pyarrow  # Required for Ray data.
sentencepiece  # Required for LLaMA tokenizer.
numpy
torch==2.8.0
transformers >= 4.37.0  # Required for Mistral.
matplotlib
plotly_express
seaborn
wandb
kaleido
ddsketch
jupyterlab
flashinfer-python==0.4.1
pillow
tiktoken
grpcio
grpcio-tools  # For protobuf compilation
protobuf>=4.21.0  # For protobuf serialization
uvicorn
fastapi
openai
tqdm
